Add IGS support to account settings
A new option, "Indiquer mon pseudo IGS", was added to the account select menu. A related command, AccountIGSCommand, was created to handle the IGS username. Furthermore, an 'IGS' field was added to the Member class and its corresponding card in Discord. In addition, a null check was added in the GameAnnouncer class to prevent potential errors.