Member list, Inovua's React DataGrid library
This commit installed babel runtime and the Inovua's React DataGrid package that provides a powerful grid component for displaying table data. In addition, two new SCSS files were added for modification and construction of the layout and style of components. The root, main and index files were adjusted to fit the implemented changes. Finally, UI components were added for navigation and displaying a list of members, utilizing the acquired grid package.