Refactor announcement methods and implement i18n service
This commit refactors the announcement methods for the RankUpAnnouncer and GameAnnouncer classes, abstracting the announcement functions for better reuse in different use-cases. Also, it introduces the I18n service to handle internationalization of user-facing strings, replacing hardcoded strings with references to a 'messages.properties' file for easier localization. A Controller for announcements was also added.