Add last check info to FFG profile reload
Updated the member class, both front and back-end, to include the information of the last check to the Federation Francaise de Go (FFG) profile. This entails refactoring the reloader button in the member list's cell render to display the `ffgLastCheck` data item. Also bumped the version in the pom.xml file.